ibm_power ohai plugin
This adds a basic ibm_power ohai plugin which parses the cpu information for
POWER8/9 machines. Some of this I will upstream into ohai itself eventually.
This also includes a useful attribute for determining whether the node is power8
or power9 which we need to know now.

I plan to add several other attributes later but this was needed initially. I
couldn't add ohai ChefSpec since we need to be using a newer ChefDK so I'll add
that as a TODO later.